Add files via upload
This dataset must be uploaded in the google drive in the correct path that is given in the code in order for the program to run and give result. This code only detects ham or spam only by processing the contents or context of the received mail, not the mail id.